WIP Updating tasks pipeline to run new parsers
